+++ src/hooks/index.ts
@@ -2,4 +2,4 @@ export { default as useCheckLogin } from './useCheckLogin';
 export { usePomodoroValue, usePomodoroActions } from './usePomodoro';
 export { default as useCurrentTodo } from './useCurrentTodo';
 export { default as useDraggableInPortal } from './useDraggableInPortal';
-export { default as useEdit } from './useEdit';
+export { useEdit, EditContextProvider } from './useEdit';
